add skip dmriprep arg

+++ hcp_pipelines/main.py
@@ -36,6 +36,7 @@ def main():
     #     help="""The local directory where the output
     #     files should be stored.""",
     # )
+    parser.add_argument("--skip_dmriprep", action="store_true")
     parser.add_argument(
         "--remove_work_dir",
         action="store_true",
@@ -139,49 +140,50 @@ def main():
         include_ses=True,
     )
 
-    # Run dmriprep
-    print("Running dmriprep...\n")
-    cmd = f"dmriprep /input /output participant -w /work_dir -s 1 \
-        --denoise_strategy {args.denoise_strategy} \
-        --participant_label {args.participant_label} \
-        --nprocs {args.n_cpus} --omp_nthreads {args.n_cpus} --mem_gb {args.mem_gb} "
-    run(cmd)
+    if not args.skip_dmriprep:
+        # Run dmriprep
+        print("Running dmriprep...\n")
+        cmd = f"dmriprep /input /output participant -w /work_dir -s 1 \
+            --denoise_strategy {args.denoise_strategy} \
+            --participant_label {args.participant_label} \
+            --nprocs {args.n_cpus} --omp_nthreads {args.n_cpus} --mem_gb {args.mem_gb} "
+        run(cmd)
 
-    # Rename files
-    input_dir = f"/input/sub-{args.participant_label}/ses-1/dwi/"
-    shutil.rmtree(input_dir)
-    shutil.copytree(
-        f"/output/sub-{args.participant_label}/ses-1/dwi/",
-        f"/input/sub-{args.participant_label}/ses-1/dwi/",
-    )
-    shutil.rmtree("/output/", ignore_errors=True)
-    input_dir = Path(input_dir)
+        # Rename files
+        input_dir = f"/input/sub-{args.participant_label}/ses-1/dwi/"
+        shutil.rmtree(input_dir)
+        shutil.copytree(
+            f"/output/sub-{args.participant_label}/ses-1/dwi/",
+            f"/input/sub-{args.participant_label}/ses-1/dwi/",
+        )
+        shutil.rmtree("/output/", ignore_errors=True)
+        input_dir = Path(input_dir)
 
-    # Make output dir per m2g spec
-    m2g_path = Path(f"/output/sub-{args.participant_label}/ses-1/dwi/preproc/")
-    m2g_path.mkdir(parents=True, exist_ok=True)
+        # Make output dir per m2g spec
+        m2g_path = Path(f"/output/sub-{args.participant_label}/ses-1/dwi/preproc/")
+        m2g_path.mkdir(parents=True, exist_ok=True)
 
-    files = list(input_dir.glob("*.*"))
-    for file in files:
-        if not "final" in file.name:
-            file.unlink()
-        else:
-            parent = file.parent
-            bids_name = f"sub-{args.participant_label}_ses-1_run-1_dwi"
-            suffix = file.suffix
-            if suffix == ".gz":
-                suffix = ".nii.gz"
+        files = list(input_dir.glob("*.*"))
+        for file in files:
+            if not "final" in file.name:
+                file.unlink()
+            else:
+                parent = file.parent
+                bids_name = f"sub-{args.participant_label}_ses-1_run-1_dwi"
+                suffix = file.suffix
+                if suffix == ".gz":
+                    suffix = ".nii.gz"
 
-            new_name = parent / (bids_name + suffix)
-            file.rename(new_name)
+                new_name = parent / (bids_name + suffix)
+                file.rename(new_name)
 
-            if suffix == ".nii.gz":
-                eddy_file = m2g_path / "eddy_corrected_data.nii.gz"
-                shutil.copyfile(str(new_name.absolute()), str(eddy_file.absolute()))
+                if suffix == ".nii.gz":
+                    eddy_file = m2g_path / "eddy_corrected_data.nii.gz"
+                    shutil.copyfile(str(new_name.absolute()), str(eddy_file.absolute()))
 
-    # Delete work dir
-    if args.remove_work_dir:
-        shutil.rmtree("/work_dir", ignore_errors=True)
+        # Delete work dir
+        if args.remove_work_dir:
+            shutil.rmtree("/work_dir", ignore_errors=True)
 
     # Run m2g
     cmd = f"m2g_bids --participant_label {args.participant_label} --session_label 1\
